Map fragment 연결
NavigationActivity -> 프래그먼트로 변경 /
프래그먼트 조합하는 combineActivity 생성
