+++ app/src/main/java/com/example/aieyes/utils/GestureManager.java
@@ -45,6 +45,8 @@ public class GestureManager {
 
             @Override
             public boolean onFling(MotionEvent e1, MotionEvent e2, float velocityX, float velocityY) {
+                if (e1 == null || e2 == null) return false;  //  방어 코드 추가
+
                 float diffX = e2.getX() - e1.getX();
                 float diffY = e2.getY() - e1.getY();
