+++ navigation/src/main/AndroidManifest.xml
@@ -1,10 +1,22 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <application>
-        <activity
-            android:name="com.example.capstone_map.feature.NavigationActivity"
 
-            android:exported="false" />
+
+        <!-- 🔥 CombinedActivity를 시작 액티비티로 설정 -->
+        <activity
+            android:name="com.example.capstone_map.feature.CombinedActivity"
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+
+        <activity
+            android:name=".feature.NavigationActivity"
+            android:exported="false"/>
+
     </application>
 
     <uses-permission android:name="android.permission.VIBRATE" />
