--- navigation/src/main/res/layout/activity_navigation.xml
@@ -1,134 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.example.capstone_map.feature.NavigationActivity">
-
-    <!-- 메인 레이아웃 -->
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <!--  카메라 프리뷰 (상단 30%) -->
-        <androidx.camera.view.PreviewView
-            android:id="@+id/cameraPreview"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHeight_percent="0.3"
-            android:background="#000000" />
-
-        <!-- 카메라 상태 표시 (카메라 위에 오버레이) -->
-        <TextView
-            android:id="@+id/txt_result"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_margin="16dp"
-            android:background="#A0000000"
-            android:padding="12dp"
-            android:text="분석을 시작하려면 버튼을 누르세요."
-            android:textColor="@android:color/white"
-            android:textSize="16sp"
-            app:layout_constraintTop_toTopOf="@id/cameraPreview"
-            app:layout_constraintStart_toStartOf="@id/cameraPreview"
-            app:layout_constraintEnd_toEndOf="@id/cameraPreview" />
-
-
-
-
-
-
-        <!-- 지도 (중앙 60%) -->
-        <LinearLayout
-            android:id="@+id/linearLayoutTmap"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            android:orientation="vertical"
-            app:layout_constraintTop_toBottomOf="@id/cameraPreview"
-            app:layout_constraintBottom_toTopOf="@id/infoPanel"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent" />
-
-        <!--  정보 패널 (하단 10%) -->
-        <LinearLayout
-            android:id="@+id/infoPanel"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:padding="12dp"
-            android:background="#DD000000"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent">
-
-            <!-- 네비게이션 안내 텍스트 -->
-            <TextView
-                android:id="@+id/navigationText"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="네비게이션 안내"
-                android:textColor="#FFFFFF"
-                android:textSize="16sp"
-                android:textStyle="bold"
-                android:gravity="center_horizontal" />
-
-            <!-- 결과 텍스트 -->
-            <TextView
-                android:id="@+id/result_text"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="4dp"
-                android:text="결과가 표시됩니다"
-                android:textColor="#BBDEFB"
-                android:textSize="14sp"
-                android:gravity="center_horizontal" />
-
-        </LinearLayout>
-
-        <!-- ✨ 제스처 감지용 투명 오버레이 (최상위) -->
-        <View
-            android:id="@+id/gestureOverlay"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:clickable="true"
-            android:focusable="true"
-            app:layout_constraintTop_toTopOf="@id/linearLayoutTmap"
-            app:layout_constraintBottom_toBottomOf="@id/linearLayoutTmap"
-            app:layout_constraintStart_toStartOf="@id/linearLayoutTmap"
-            app:layout_constraintEnd_toEndOf="@id/linearLayoutTmap"
-            android:background="#00000000" />
-
-
-
-        <!-- 로딩 표시 (중앙) -->
-        <ProgressBar
-            android:id="@+id/progressBar"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:visibility="gone"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <!-- 🔑 버튼을 FrameLayout 최상위로 (오버레이 위!) -->
-    <Button
-        android:id="@+id/btn_toggle_analysis"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="top|center_horizontal"
-        android:layout_marginTop="260dp"
-        android:text="분석 시작"
-        android:elevation="8dp" />
-
-
-
-
-</FrameLayout>
